# mysite/products/openapi/products-api.yaml
openapi: 3.0.3
info:
  title: Products and Categories API
  version: "1.0.0"
  description: >
    Contrat OpenAPI pour la gestion des modeles Django Product et Category.
    Cette specification sert de base avant et pendant l'implementation des endpoints.
servers:
  - url: /
...
paths:
...
  /api/products/{product_id}/:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      summary: Recuperer un produit
      operationId: getProduct
      responses:
        "200":
          description: Produit detaille
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Products]
      summary: Mettre a jour partiellement un produit
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
            example:
              stock: 8
              is_active: true
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProductPatchMultipart"
      responses:
        "200":
          description: Produit mis a jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
    delete:
      tags: [Products]
      summary: Supprimer un produit
      operationId: deleteProduct
      responses:
        "204":
          description: Produit supprime
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
...
    ProductId:
      in: path
      name: product_id
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Donnees invalides
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    UnsupportedMediaType:
      description: Type de contenu non supporte
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    CategorySummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    Product:
      type: object
      required:
        [id, category, sku, name, slug, description, image_url, price, stock, is_active, created_at, updated_at]
      properties:
        id:
          type: integer
        category:
          $ref: "#/components/schemas/CategorySummary"
        sku:
          type: string
          maxLength: 50
        name:
          type: string
          maxLength: 200
        slug:
          type: string
        description:
          type: string
        image_url:
          type: string
          nullable: true
          format: uri
        price:
          type: string
          example: "149.90"
        stock:
          type: integer
          minimum: 0
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProductPatch:
      type: object
      properties:
        category_id:
          type: integer
        sku:
          type: string
          maxLength: 50
        name:
          type: string
          maxLength: 200
        slug:
          type: string
        description:
          type: string
        price:
          type: string
          example: "149.90"
        stock:
          type: integer
          minimum: 0
        is_active:
          type: boolean
        clear_image:
          type: boolean
          description: Permet de supprimer l'image courante
    ProductPatchMultipart:
      allOf:
        - $ref: "#/components/schemas/ProductPatch"
        - type: object
          properties:
            image:
              type: string
              format: binary
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationFieldError:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationFieldError"